pyproject.toml
[project]
name = "web3-chatbot-agent"
version = "0.1.0"
description = "Web3 Agent Chatbot - An AI-powered assistant for blockchain interactions"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chainlit>=2.4.400",
    "python-dotenv>=1.0.0",
    "web3>=7.0.0",
    "openai>=1.12.0",
    "openai-agents>=0.0.12",
    "logfire>=3.14.1",
    "py-solc-x>=2.0.3",
]
